包裹的查询

+++ src/main/java/scc1/java_project/deliver_server/POJO/SimpleOrderInfoBar.java
@@ -6,7 +6,7 @@ import java.sql.Timestamp;
 
 @Data
 public class SimpleOrderInfoBar {
-    private long orderId;
+    private int orderId;
     private Timestamp orderCreateTime;
     private String receiveName;
     private String orderStatus;
